Add CLI, smoke tests, and fix polling timeout logic
- auteur/cli.py + __main__.py: `python -m auteur run/resume/status`
  with cost summary printed on completion
- scripts/smoke/: three scripts (qwen, image, i2v) plus run_all.sh
  to validate API key and model IDs before a full pipeline run
- Fix poll_until_done to use wall-clock time (time.monotonic) instead
  of accumulated interval sum — the previous approach hung forever
  when poll_interval=0
- Widen poll_interval/timeout to float; adjust config defaults to match

+++ auteur/config.py
@@ -24,8 +24,8 @@ class Settings(BaseSettings):
     parallel_shot_limit: int = 3
 
     # Job polling
-    job_poll_interval: int = 10   # seconds between polls
-    job_timeout: int = 600        # max seconds to wait for one job
+    job_poll_interval: float = 10.0   # seconds between polls
+    job_timeout: float = 600.0        # max seconds to wait for one job
 
     # Storage
     project_dir: str = "./projects"
